solved the issues in EditPage, Updated SnackBar, BibleDropdown etc
