+++ src/components/EditProject/BibleBookTable.js
@@ -8,30 +8,6 @@ import TableContainer from '@material-ui/core/TableContainer';
 import TableHead from '@material-ui/core/TableHead';
 import TableRow from '@material-ui/core/TableRow';
 
-const columns = [{ id: 'name', label: 'Books Assigned' }];
-
-function createData(name, code, population, size) {
-  const density = population / size;
-  return { name, code, population, size, density };
-}
-
-const rows = [
-  createData('India', 'IN', 1324171354, 3287263),
-  createData('China', 'CN', 1403500365, 9596961),
-  createData('Italy', 'IT', 60483973, 301340),
-  createData('United States', 'US', 327167434, 9833520),
-  createData('Canada', 'CA', 37602103, 9984670),
-  createData('Australia', 'AU', 25475400, 7692024),
-  createData('Germany', 'DE', 83019200, 357578),
-  createData('Ireland', 'IE', 4857000, 70273),
-  createData('Mexico', 'MX', 126577691, 1972550),
-  createData('Japan', 'JP', 126317000, 377973),
-  createData('France', 'FR', 67022000, 640679),
-  createData('United Kingdom', 'GB', 67545757, 242495),
-  createData('Russia', 'RU', 146793744, 17098246),
-  createData('Nigeria', 'NG', 200962417, 923768),
-  createData('Brazil', 'BR', 210147125, 8515767),
-];
 
 const useStyles = makeStyles({
   root: {
@@ -63,20 +39,7 @@ const useStyles = makeStyles({
 
 export default function StickyHeadTable(props) {
   const classes = useStyles();
-  const [page, setPage] = React.useState(0);
-  const [rowsPerPage, setRowsPerPage] = React.useState(10);
-
   const bookAssigned = props.projectData.metaData.books;
-
-  const handleChangePage = (event, newPage) => {
-    setPage(newPage);
-  };
-
-  const handleChangeRowsPerPage = (event) => {
-    setRowsPerPage(+event.target.value);
-    setPage(0);
-  };
-  console.log('aaaaaaaaaaassssssssss', bookAssigned);
 
   return (
     <Paper className={classes.root}>
@@ -92,7 +55,6 @@ export default function StickyHeadTable(props) {
           <TableBody>
             {bookAssigned.map((row) => {
               const index = bookAssigned.indexOf(row);
-              //   console.log('ppppppppppppppppppppp',index)
               return (
                 <TableRow key={index}>
                   <TableCell className={classes.tableBody} key={index}>
