+++ src/components/SnackBar/SnackBar.js
@@ -12,6 +12,7 @@ export default function SnackBar(props) {
 
   return (
     <Snackbar
+      style={{ height: "100%" }}
       open={responseStatus[0]}
       autoHideDuration={6000}
       onClose={props.handleClose}
