+++ src/components/UploadSource/UploadSource.md
@@ -8,6 +8,8 @@ import Button from '@material-ui/core/Button';
 import Grid from '@material-ui/core/Grid';
 import { API } from '../../store/api';
 import SnackBar from '../SnackBar/SnackBar.js';
+const [loading, setLoading] = React.useState('');
+import CircularProgress from '@material-ui/core/CircularProgress';
 
 const [selectedFiles, setSelectedFiles] = React.useState('');
 const [responseStatus, setResponseStatus] = React.useState([]);
@@ -16,72 +18,8 @@ const handleClose = () => {
   setResponseStatus([false]);
 };
 
-const [projectData, setProjectData] = React.useState({
-  projectId: 100000,
-  projectName: 'Mission Agape',
-  sourceLanguage: {
-    languageId: 100037,
-    language: 'English',
-    code: 'en',
-    scriptDirection: 'left-to-right',
-    metaData: {
-      region: 'United Kingdom, Europe',
-      'alternate-names': [
-        'Anglit',
-        'Kiingereza',
-        'Gustavia English',
-        'Samaná English',
-        'Saint Lucian English',
-        'Noongar',
-        'Noonga',
-        'Newcastle Northumber',
-        'Neo-Nyungar (Noogar)',
-        'Glaswegian',
-        'Brummy',
-        'Birmingham (Brummie)',
-        'Bay Islands English',
-        'Australian Standard English',
-        'Aboriginal English',
-        'African American Vernacular English (AAVE)',
-      ],
-      'suppress-script': 'Latn',
-      'is-gateway-language': true,
-    },
-  },
-  targetLanguage: {
-    languageId: 100057,
-    language: 'Hindi',
-    code: 'hi',
-    scriptDirection: 'left-to-right',
-    metaData: {
-      region: 'India, Asia',
-      'alternate-names': [
-        'Khadi Boli',
-        'Khari Boli',
-        'Dakhini',
-        'Hindi-Urdu',
-        'Khariboli',
-      ],
-      'suppress-script': 'Deva',
-      'is-gateway-language': true,
-    },
-  },
-  documentFormat: 'usfm',
-  users: [
-    {
-      project_id: 100000,
-      userId: 10101,
-      userRole: 'owner',
-      metaData: null,
-      active: true,
-    },
-  ],
-  metaData: {
-    books: ['mat', 'luk', 'jhn', '3jn'],
-    useDataForLearning: true,
-  },
-  active: true,
-});
+const projectId = 100000;
+  
 
 const clearState = () => {
   setSelectedFiles('');
@@ -104,6 +42,8 @@ const readFileAsText = (file) => {
 };
 
 const loadText = () => {
+  if(selectedFiles.length > 0){
+  setLoading(true);
   let files = selectedFiles;
   let readers = [];
   for (let i = 0; i < files.length; i++) {
@@ -112,36 +52,38 @@ const loadText = () => {
 
   Promise.all(readers).then((values) => {
     const data = {
-      projectId: projectData.projectId,
+      projectId: projectId,
       uploadedBooks: values,
     };
     API.put('autographa/projects', data)
       .then(function (response) {
         setResponseStatus([true, 'success', response.data.message]);
-        console.log('ssssssss', response.data.message);
+        setLoading(false);
       })
       .catch((error) => {
         setResponseStatus([true, 'error', 'failed']);
-        console.log('eeeeeeeeeeeeeee', error);
+        setLoading(false);
       });
-  });
+  });}else{
+    setResponseStatus([true, 'error', 'Please select any files']);
+  }
 
   clearState();
 };
 <>
   <Grid container direction='row'>
     <SnackBar responseStatus={responseStatus} handleClose={handleClose} />
-    <UploadSource projectData={projectData} onChange={setSelectedFiles} />
+    <UploadSource onChange={setSelectedFiles} />
     <Button
       color='primary'
       size='small'
       variant='contained'
       component='span'
-      disabled={!selectedFiles}
       onClick={loadText}
     >
       Upload
     </Button>
+    {loading && <CircularProgress size='1.5rem' />}
   </Grid>
 </>;
 ```
